Bye, Travis...
The once very great Travis CI service, due to an acquisition by
"Idera", is now no longer a viable option for open source use.  The
company is not responding to requests for free plans for open source
projects (despite claiming they offer them), and the way in which a
metered/billed system was suddenly rolled out did not give warning to
allow people to make other preparations.  It has numerous glitches that
would add even more wariness to being willing to give the people
involved one's credit card as an answer to the problems.

This removes .travis.yml and related files, in order to begin building
new CI answers from scratch.