Eliminate the Pthread Build of Emscripten
This retains the relevant code in load-r3.js (in case something
analogous comes up in the future).  But it makes a commitment to not
using pthreads for the implementation of interpreter suspension for
interactivity.  `Asyncify` is used for now to unwind stacks (at the
cost of bloating the build size), with the methods of the "Stackless"
branch promising a *better* answer than pthreads.

See rationale here:

https://forum.rebol.info/t/pros-and-cons-of-the-pthread-web-build/1425