Fix fail() on PAR() parameter arguments
To make raising errors easier, it is possible for natives to fail on
the argument's cell in the paramlist, to indicate that the parameter
with the invalid value should be called out by name vs. just erorring
on a plain "bad value".

The facility was broken when paramlists were rethought to be shared
between functions.  This updates it for the new rules.