Fixes for http protocol "chunked mode"
The http chunked mode must not be used that often in practice, because
it looks to have been broken for a while.

In addition to some changes that broke what would have worked before,
it assumes the protocol also assumed that the hex chunk sizes had an
even number of digits that DEBASE with /BASE 16 would accept.  That's
not always true, apparently.

So this patches those issues enough to load a chunked page that was
not loading before.