SOURCE fix for specialized functions
Specialized functions no longer need to store the specialized FRAME!
in their ACT_DETAILS(), they can just use the exemplar.  (This allows
specializations to be more cheaply allocated.)

However, SOURCE was still using the old expectation and looking in
the details instead of the exemplar.  This was covered up by another
issue in a failing test...this fixes both issues.