Default REDUCE predicate discards NULL vs. erroring
Ren-C offers the advantage of a "true NULL" type that cannot be mistaken
for a value desired to be intentionally put into a block.  Several
places leverage this to allow the NULL intent to signal vaporization,
e.g. COMPOSE.

One place it was held off on exploiting this vanishing property was
in REDUCE.  Instead, it chose to error by default.  This retained a
property of historical Rebol's REDUCE where evaluating N expressions
would produce N results...which was considered a comfortable idea
e.g. in block-based SET assignments:

    ; If NULL dissolves, an assignment like this could shuffle the
    ; intended alignments of expressions with variables
    ;
    set [a b c] [expr1 expr2 expr3]

However, many common cases make more sense to use the dissolving:

    >> append [<a> <b>] reduce [<c> if false [<d>]]
    == [<a> <b> <c>]

This commit changes to that behavior as the new default, instead of
raising an error, due to the generally popular opinion that this is
the more common idiom.

However...REDUCE is now parameterizable by a "predicate" that can
operate on the values as they are collected, which provides
configurability for either erroring or transformations:

    >> append [<a> <b>] reduce .non.null [<c> if false [<d>]]
    == [<a> <b> <c>]
    ** Error: NON failed with argument of type NULL

    >> append [<a> <b>] reduce .try [<c> if false [<d>]]
    == [<a> <b> <c> _]