LOAD of code always BLOCK!, LOAD-VALUE for 1 item
The historical behavior of LOAD included a twist that created a
variability that made an already complex function more complicated:

    rebol2> load "1 2"
    == [1 2]

    rebol2> load "1"
    == 1

The "/ALL" refinement could be used to stop the extraction of single
elements from a block:

    rebol2> load/all "1"
    == [1]

This commit removes the /ALL refinement and makes LOAD/ALL the only
behavior offered by LOAD.  However, it gives LOAD-VALUE as a variant
which will ensure only a single value is loaded:

    >> load-value "1"
    == 1

    >> load-value "1 2"
    ** Error: LOAD-VALUE got length 2 block, not 1

This should offer a suitable workaround for those wishing to load and
bind single values.  Better names or ideas solicited, but it's a good
change to make the main LOAD more consistent.

An adapter is added to %redbol.reb for a compatibility version that
adds the /ALL refinement back via AUGMENT and implements the old
behavior:

    load: enclose (augment :load [/all]) func [
        f <local> try-one-item
    ][
        try-one-item: not f/all
        result: do f  ; now always BLOCK! if LOADing Rebol code

        if try-one-item and (block? result) and (length of result = 1) [
            return first result  ; "1" loads as `[1]`, change it to `1`
        ]
        return result  ; "1 2" loads as `[1 2]`, leave it that way
    ]