Fix release build issue with unreadable voids
Due to the desire to manifest an unreadable void as something that
won't crash the system if the user somehow sees it in the release
build, it uses a real interned string...but this string wasn't available
early enough in the boot for some uses.

This adds it to the special symbols loaded before the boot code.