Cleanup stray AND with BLOCK! as right hand side
It was noticed that the Q console shortcut did not work.  This was
because it was using the AND operation from when it took a BLOCK!
on the right hand side.

That predated the concept that quoting GROUP!s and running them
conditionally was a logical behavior at callsites if the context
called for it (e.g. not running a GROUP! in the branch slot of a
conditional if the conditional was not true).

A search found several other places that hadn't been fixed, this
updates them.