Mothball the SECURE Extension
While the features of SECURE are certainly desirable, it is more of
an OS/runtime question than a language design question.  It's also
difficult to make claims using the word "security" in this day and
age, especially with the language model in use.

Rather than keep the unfinished code around, this leaves it in the
Git history to perhaps be looked at later for ideas.

https://forum.rebol.info/t/secure-is-disabled/890