Tidy up REQUEST-FILE
The REQUEST-FILE feature in the View module is a GUI feature that is
actually relatively low-impact on Windows, since it is a common
dialog provided by the OS.  The code is actually a pretty good
demonstration of how to use the libRebol API.

This brings it up to date, leaning less on core APIs and more fully
on libRebol features.  It also corrects a bug with the file filtering
that relates to the illegality of 0 bytes in TEXT! strings.