Fix ESC in console, new-line in empty REDUCE
Due to REDUCE vaporizing NULL, there was an empty block occuring
where it did not before.

This exposed a bug in how the new-line flag was being propagated from
the original block in a REDUCE to the generated block, which triggered
an assert because it tried to set it on cells even when no material
was added to the reduce.  (This problem existed before, it just became
more likely to happen with vaporizing being the default.)