Preliminary implementation of "medium escapability"
The historical idea of "soft quoting" has accomodated the idea of
functions which quote but are willing to be deferent to certain enfix
operations, such as how branches are willing to allow lambdas:

    if true [...] then x -> [...]
    if true [...] then '[block by value]

Without quoting, then the block by value parameter would not be
received with its quote on it (so it would not be known that it was
meant to be by value).  Yet if quoting were non-negotiable, the case
of the THEN would get `x` and not let the lambda run.  So soft quoting
is willing to defer its claim on `x` to the enfix quoting operation ->

This introduces the new idea of "medium escapability" which is still
escapable via GET-WORD!/GET-PATH!/GET-GROUP!, but does not delegate to
enfix operations in this way, either leftward or rightward.

!!! The change is not implemented fully, e.g. for variadics--which
are in need of rewrite and review in any case.  But it seems clear
that this form is needed as a distinction; so it goes ahead and puts
it on the map.