Patch for quoted refinements
Quoted refinements were not working correctly, due to issues with
recognizing the `:/foo` and `'/foo` patterns, which are not plain PATH!

The userspace FUNCTION implementation was missing these cases, and
the GET-PATH! case was also being missed in the native implementation.

This works around it, but a cleaner approach is needed.