Handle quoted nulls followed by comments
The scanner handles quoting by holding the apostrophes as a number
of `quotes_pending` and then applies that quoting to whatever an
ordinary scan produces.  This has to watch out for if nothing is going
to come next (indicating a quoted null).  So it was checking for spaces
or ']' or ')'.

However, it wasn't checking for comments.  So `';` would leave the
quote pending, and then apply to whatever token came next.  If that
was the end of the input, it would assert.