Eliminate ARRAY_FLAG_NULLEDS_LEGAL
This was a debug-only flag for checking when special arrays were
permitted to contain nulls.  For a time this could apply to reified
evaluation streams--though that hadn't been thought completely through,
as such arrays might leak to the user.  With the advent of generic
quoting, the idea was completely eliminated.

This means the only legal places to find nulled cells is in contexts,
which are able to hold the nulled state.  But there's already a
flag for that (ARRAY_FLAG_IS_VARLIST), so that is checked.