Split out VAL_ACTION_BINDING() from VAL_BINDING()
Previously, the "extra" slot was used for binding information, which
could mean any number of things depending on if it were an array, or
an action, or a word.  If it was an array it would be less of a
binding and instead carry a "specifier" (if the array was in a
REBVAL on the stack) or relative information (if it was itself in
an array for the body of a function).

With specifiers about to become more complex, there is an idea that
stack cells for words may need to store specifiers.  Because of the
very heterogeneous notions of what a "binding" is, it makes more sense
to rally the common use of the field for specifiers.

As a first step, this breaks out the binding of actions to its own
extractor and initialization.  This permits it to be narrowed to where
action bindings are limited to contexts.