Eliminate VAL_CONTEXT_VAR()/VAL_CONTEXT_KEY()
As a convenience, it was once seen as desirable to offer macros so
that you instead of writing:

    CTX_VAR(VAL_CONTEXT(context), n)

You could write:

    VAL_CONTEXT_VAR(context, n)

It's not too much of an abbreviation.  But in addition, VAL_CONTEXT()
now does more work as a kind of "gatekeeper" on noticing when a
context is inaccessible due to being freed, and can raise an error.
This makes it desirable to avoid redoing that check.

Furthermore, there are aspects to which having a full value comes into
play as more meaningful, e.g. by taking into account the REBVAL's
phase information for frames to saying whether a key is visible or not.
This wasn't being covered by simple macros like VAL_CONTEXT_KEY, which
misses much of the point of using a full value cell instead of a node.

This scraps the macros altogether, as there weren't all that many uses,
and documents the issues for why they no longer exist--putting focus
on the primitives that actually show a difference when given values.