Reduce states for label handling in FRAME!
The ACTION! label handling was not written to deal with null labels.
Instead, it would just leave an action looking like the archetype,
with the specialty information instead of a string in the slot.  This
was interpreted as anonymous.

FRAME! labels instead would store nulls.  As frames had to handle the
archetypal unlabeled case either way, this meant there were a total
of three states instead of two.

This makes frames use the same concept as actions, but changes the
interface to do handle null labels as filling in the archetypal
information.