Case usage updates (FAIL, PANIC, headers)
A short-lived experiment tried capitalizing FAIL and PANIC in order
to make them stand out more.  It would be possible to continue this
idea in a case-sensitive binding world by creating aliases, but the
idea isn't all that great to start with so this just removes it--at
least for now.

More significantly for running in a case-sensitive binding world, this
changes the sysobj.r example module header fields to be titlecase so
that it works with the standard capitalized header fields that have
been traditionally used in Rebol.