Get rid of SELF in the internals of the system
SELF had been phased out of several internal layers of the system, so
there was nothing "built in" about the word that would break its use
for other applications.

So it was essentially removed, yet it was hanging around in objects
and contaminating some of the binding and object definition mechanics.
This goes ahead and kills it off completely.

Getting at an object can be done with asking for the binding of any
of its bound keys.  So if one wanted to write a SELF that would work
for inherited cases, you should say:

    self: method [] [binding of 'self]