Move Primary Binding Index into QUOTED! cell
Though there is a trick for getting small levels of escaping to fit
into a cell, higher levels require making a new allocation.  The new
allocation is a distinct cell type...that must be able to present as
either bindable or unbindable (depending on what type is quoted).

As a first cut of dealing with that situation, all QUOTED! cells
reported that they were bindable, and the binding was stored in the
outer container.  However, they had to synchronize with the binding
information of the inner word cell, because that's where the index
was stored...which conflicted with the quoting depth.

This rethinks that a bit to give up some bits of the bind index in
order to put it in the outer container merged with the depth.  This
means all binding information can fit in the container, which means
the escaped node doesn't have to be recreated for new bindings.