Remove DEBUG_BINDING_NAME_MATCH
The debug option for checking that bindings name matches would no
longer be applicable, as we leverage the assumption of the names
matching to reclaim bits in bound words for usage.

Also changes Get_Word_Context() to return an index, so that it is not
presumed that the word's stored index is the one that matched in the
bind (which it wouldn't be for virtual binding's case).