Quoting workaround for NULL in instruction array
This is a quick workaround to try and deal with a problem that was
hit while running cases like:

    >> recycle/torture
    >> until .not.quoted? [null]

This was causing an assertion because a NULL was being rebQ()'d, and
this transiently creates an "instruction" array that is not covered
by the existing exemptions for NULLs being legal in arrays.

Rather than add the instruction to the exemption list, this tweaks
the rebQUOTING to always hold quoted elements and then quotes them
one level less.

This isn't really very well thought out, and the whole attempt to
generalize rebQ() and rebU() needs work.  This quick fix is just to
keep things running for the common case, and review that later.