Terminate virtual bind chains with the frame specifier
The original conception of virtual binding worried that composing
together virtual binds would be common enough that storing a single
resolving frame in the tail of the chain would cause long chains to
be repeated several times, changing only that specifier.  So it was
thought that caching the specifier in the head of the chain would be
a better idea, so two virtual bind chains needed to be glued together
they could not worry about an incompatible specifier.

But looking at cases in practice, this may not be that common.  The
more valuable concept for using the space in the head of the chain
is likely as a way of linking similar chains together so that they
can find each other instead of generating a new one each time a
similar chain occurs.