Make PRINT write to console.log() in JavaScript boot
It's difficult to debug aspects of the mezzanine portion of the boot
process without some I/O.  This happens before the JS build hooks
print with the ReplPad.

Since emscripten offers printf(), this changes the boot WRITE-STDOUT
from sensing its availability based on being a debug build, and instead
the DEBUG_STDIO_OK switch, which can be turned on in emscripten
"release" builds.