Stop Typeset_To_Array() from crashing on custom types
The idea behind experimental custom datatypes is to sacrifice one of
the 4 platform pointer slots in the value cell for a pointer to a
method table for that type.

This experiment ran up against the hardcoded notion of "at most 64
built in types" of historical Rebol.  An obvious problem is that all
the custom types are considered equivalent in that scheme.

A crash was triggered when trying to display a typeset containing
one of these seminal custom types.  This just throws in an interem
rendering of ~CUSTOM!~.  However, all of this code is slated to be
completely rethought:

https://forum.rebol.info/t/the-typeset-representation-problem/1300