Time constructor with nanoseconds for NOW/TIME/PRECISE
The MAKE TIME! and MAKE DATE! constructions in historical Rebol
predated the /PRECISE refinement to NOW, which added sub-second
precision.  Only the C code inside of NOW was able to build times and
dates with fractional components (or time zones).

The targeted MAKE-DATE-YMDSNZ function was provided to implement NOW
in the web build.  However, this did not account for cases where the
NOW/TIME/PRECISE variation was used.

Rather than make a date and extract the time from it, this goes ahead
and adds a parallel MAKE-TIME-SN function...as a further placeholder
for the desired granularity.