Zero uninitialized bytes in cells
In cells that do not use all of their bits (such as BLANK!) it may
seem that you can leave the 3 unused cell slots uninitialized, with
whatever memory they contianed before.  Only the header would need
to be given valid bits.

As it turns out, this leads to undefined behavior when cells are
copied.  Even if the only access is to move the uninitialized data
from one blank cell to another, in *theory* a C compiler could use
some bit patterns as "trap representations".

While it might seem okay to ignore this as an unlikely case, there's
at least one good reason to pander to it (besides being on the right
side of the standard).  The reason is that there is a compiler warning
that is frequently informative that triggers if you don't do the
assignment.  Disabling this warning would lose valuable information.

This replaces previous ad-hoc approaches to the problem with a
philosophy that ZERO_UNUSED_CELL_FIELDS should be heeded to make sure
all cells are initialized with bits in release builds.

However...debug builds use the unused fields to poke in additional
information that can prove valuable for debugging.  So this commit
does some rearrangement to improve the interoperability between that
tracking and this flag.

This includes some reform which simplifies the cell initialization
process to where fewer routines pass in file and line information to
helpers in the stack.  Many of these helpers predated the existence
of DEBUG_TRACK_EXTEND_CELLS, which is now the go-to method for tricky
situations that are hard to debug and need more information.  Now
that the file and line information can come from cells themselves
even if they carry payloads, the helpers can be much simpler.

While this doesn't exhaustively push through the change to every
Init_XXX() definition just yet, it is a large simplification that
will lead to improved results, and is better than ad-hoc hacks to
work around the compiler warning at callsites.