Arrays, Binaries, and Strings derive from Series
While REBARR, REBBIN, and REBSTR all technically are just different
types of REBSER, they did not initially use derivation in the C++
build to distinguish them.

This mostly came from feeling that there were routines that could
accept a REBSER* which should not accept a REBARR*.  Also some methods
like SER_LEN() would be different for strings and binaries.  So there
was enforced casting.

But since casting is dangerous, the casts were changed to constructs
that could be checked more in the C++ build.  This gave rise to things
like SER(), which would have to be used with things like strings or
arrays if you wanted to access any series properties (like header bits)

However, most of the concerns pertaining to allowing the sublasses to
be passed around faded with time.  (e.g. SER_LEN() doesn't exist, only
SER_USED()).  The code is clearer and simpler with the C build using
the same type (Reb_Series) for all, and the C++ build using derived
classes.

With the burden lowered on using distinct types, this goes ahead
and makes binary another separate class...with strings derived from
binaries.