Eliminate separate Manage_Array() routines
Because an array is-a series, there's no need for a helper that manages
arrays, specifically.  The one thing they might do would be to cast
the result to an array for chaining, but as the comments described
it wasn't doing that since usually you don't want to use the result,
and that caused a compiler warning.