Cull the `unstable` annotation experiment
The unstable annotation was an idea of using the `volatile` qualifier
in certain debug builds to denote a slot that was potentially
movable, and hence should not exist across evaluations.

At first this was just applied to data stack pointers, but then it
seemed applicable to array pointers as well.

What this sort of ended up doing was spreading to mark the entire
codebase as "contaminated due to being C".

Making it particularly difficult to use is the practice of combining
initialization with returning, e.g.

    REBVAL *Init_XXX(RELVAL *out, ...) {
        out->... = ...;
        return cast(REBVAL*, out);
    }

The problem is that you want the unstable-ness of the output to
match that of the input, which is laborious to express in C++ with
templates, and even moreso in "C-that-can-compile-as-C++".

This is being shelved to make room for approaches that are more likely
to be effective.