Introduce DEBUG_EXTANT_STACK_POINTERS
This is a replacement for the desired functionality that the `unstable`
annotation did not produce (despite making a big mess.)

It uses a C++ wrapper class for pointers into the data stack, that
counts to make sure the extant pointers held by the user is zero when
pushing or popping, or running an evaluation.

For ergonomic reasons, the class will implicitly coerce to REBVAL*.
So it's up to callers to follow the rule of not storing the result
of DS_AT or DS_TOP in local REBVAL* variables.

Since this is a C++ class that relies on constructors and destructors,
it's technically undefined behavior if a longjmp due to fail crosses
the class.  Generally speaking this just means the destructor will
not run; and there's a workaround for it, since it's debug info
and not needed to get any particular functionality.