Introduce ARRAY_FLAG_IS_KEYLIST
Historically it wasn't considered important to specially mark keylists
of contexts, since they do not have any special GC behavior.  But
beyond being useful debugging information, there's now a concept to
decay references to expired frames to point to just the keylist.  This
would save on memory, while still maintaining a memory of the fact
that the word has been bound to something that has expired.

An additional twist is put in by making STRING_FLAG_IS_SYMBOL and
ARRAY_FLAG_IS_KEYLIST the same bit.  This anticipates the possibility
of being able to use a REBSTR* node itself to represent a single
length keylist for objects, which would be useful.