Streamline DEBUG_CHECK_CASTS
When strings, arrays, and binaries were made to be derived from
the series base class in the C++ build, this reduced the demands
on the SER() function/macro.  That made it possible to streamline
all the DEBUG_CHECK_CASTS macros somewhat to no longer need to
be used with their own types (e.g. SER() on a REBSER*)