DEBUG_CHECK_CASTS and DEBUG_EXTANT_STACK_POINTERS fix
The combination of these debug flags ran up against the desire of
smart pointer classes trying to be transparent.  This mitigates the
problem by adding an implicit constructor for REBCEL(const*) from
a STKVAL(*).

The intersection of checks means there still has to be a bit of
intervention with casting to distinguish some situations, but only
one such case was necessary so far... to write VAL_DECIMAL(VAL(DS_TOP))
instead of simply VAL_DECIMAL(DS_TOP).

This exposed some superfluous overloads that were deleting functions
that were not possible to run without the overload.