Patch ASSERT_ARRAY() for paramlists
The LET hack led to paramlists needing to signal expandability even
though they are ultimately fixed size.  This means it mucked with the
cell formatting, which was supposed to protect against overage use
in the allocation of fixed size arrays.

While that issue remains in flux, soften the assert to exclude
paramlists...leaving enough information to put it back when things
get figured out.