Separate partials into stub on function interface
When specialization was first being introduced, there was an attempt
to fit all of the information for the specialization into the FRAME!
that was being used as the example values.  This included the idea
of encoding partial specializations--which involved ordering.

The ordering of partial specializations made use of unused cells in
the exemplar.  This was an interesting though somewhat convoluted
concept, as the exemplar had to be searched to find them...and they
had to be pushed in a certain priority order to fulfill efficiently.

Since that first take on specialization, many things have changed:
notably that partial refinement specialization now only exists in
the form of naming the refinement in an order (because refinements
are only one argument, and they -are- the argument).

Also, the structure of actions has factored out the interface
information separately from the identity.  That interface has a
fairly clear point of control, which can handle more complex structures
in a localized way.

This goes ahead and does that by making actions able to have their
interface be either a paramlist, an exemplar, or a list of partials
which points to a paramlist or exemplar.

The resulting simplifications are significant; making it possible to
quickly get the partial specializations without having to make a
separate pass.