Function REORDER capability
REORDER allows you to create a variation of a function that uses the
same underlying frame format, but reorders the parameters.  For
instance, APPEND usually expects the series to append to as the
first item:

     >> append [a b c] <item>
     == [a b c <item>]

But a REORDER takes a block of parameters in the order you wish:

     >> append-value-first: reorder :append [value series]

     >> append-value-first <item> [a b c]
     == [a b c <item>]

It's currently necessary to specify all the required parameters in a
reordering.  However, optional parameters may be mentioned as well:

     >> append-val-dup-ser: reorder :append [value dup series]

     >> append-val-dup-ser <item> 3 [a b c]
     == [a b c <item> <item> <item>]

This feature effectively exposes a more ergonomic form of the reorder
that is possible using refinements in path dispatch.  The same mechanic
of applying a second pass over the frame (using indices accrued during
the first pass) is used to achieve it:

     >> append/series <item> [a b c]  ; use series parameter on 2nd pass
     == [a b c <item>]

But `:append/dup/series` is not very intuitive for getting the order
of [value dup series] (and gets more counterintuitive the more normal
parameters a function has).

Calling the argument "reorderee" seems a bit much to expose to the user,
so this pulls back on that by using those names internall and takes
them off of ADAPT so it uses `action` instead of "adaptee" as the
function parameter name.