Don't use value-based IS_END on keys, rename f->param
When keys are changed to not use cells, they will not use the IS_END()
operator.  There will have to be a nullptr terminator or enumerations
stop at a tail.

(One thing making nullptr termination less desirable is that it would
not allow keylists to be (easily) represented by a single REBSTR*)