Make it possible to Recycle() during Startup_Natives()
Much of the verification of the system's state is now tied into the
traversal done by Recycle...so a heavy debugging mode is to ask to
recycle on every step, to catch the moment of corruption.

This wasn't allowed to happen while the natives list was empty or
partially filled, so a check was added to see if a given slot for
a native was null or not.  However, the test which does a shutdown
and then starts back up again noticed that this isn't viable when
it cannot trust the zero initialization of the list, so a shutdown
function was needed.

So long as natives were being tinkered with, this moves the startup
and shutdown into the file %c-native.c  ... and the generic startup
code also.