Reuse interface in typechecker and DOES
There were a couple of places that build parameter lists by hand in
the core.  Now that they are more complicated structures--as well as
fully shareable--it makes more sense to twist their Make_Action() to
reuse an existing spec.

In the case of typecheckers, they reuse the paramlist for NULL?

For DOES of a BLOCK!, a new stub function that takes no argument and
can return any type was made...though no implementation provided for
it yet.