Bootable system with keylists 1/4 the size
This changes the interim keylist implementation using unbound WORD!
cells to the more compact representation using just a REBSER of
REBSTR*.

In doing so, it also removes the idea of a "rootkey" so that the
keylists may be as few as 1 or 0 items.  This would allow compaction
to using simply a lone REBSTR* as a keylist (though it would need the
key enumeration to be terminated by means of its length or a tail
pointer, which the system isn't quite set up for yet).