Kill off RESKINNED
RESKINNED was an early test of how to create a variation of a function
that had different type checking or parameter quoting/gathering bits.

It didn't work very well initially, and it became more broken over time.
The recent changes to paramlists and keylists completely break it.
This just removes it, and puts the code for the tests on the forum
for quick reference in case it's interesting to look at:

https://forum.rebol.info/t/1223/3