Rename "special" to "param", introduce REBVAR/REBPAR
With the changeover of `param` to `key` complete, this now turns the
former `special` into `param`...because it unifies the specialization
data along with the typeset information.  If a slot is specialized or
local, then it has no parameter description information for that phase.

This also makes subclasses of REBVAL to make it easier to do things
like compile-time type check how the MARKED cell flag is applied.