Don't repeat first param finding in invocation frame
The initial implementation of creating a frame for invoking a function
was for MATCH.  (This functionality has been generalized in the
REFRAMER concept.)

It was first thought that since the code for making an invocation
frame walked through the parameter list to fulfill it, that it could
also remember the first unspecialized parameter.  This is true, but
that's not what was happening...it was doing a separate walk, and not
using any of the current logic.

Since the optimization wasn't an optimization, this removes it.