Add a primitive AS FRAME! for action and vice-versa
With exemplars being fully formed as contexts, with separated keys
and values, this opens up the idea of viewing an action "as a frame"
and being able to access whatever it uses to hold parameter
information.

This makes the initial step of adding those cases to AS.  It replaces
the notion of asking for an array of typesets, which is of little
help when you don't know what the ordering of that array is (since
there is a REORDER facility).  AS FRAME! action gives you the types
such that you can access them by key.