Eliminate "typeset strings"
The recent commits made the appearance of a string in a "PARAM!" value
a temporary state, only used while building up parameter lists on
the stack.  It was then extracted into a keylist, with the paramter
holding the information nulled out.

This takes a next step by collecting the symbol in a separate WORD!
on the stack, so there is never a state where params hold symbols.
That eliminates the VAL_TYPESET_STRING() and node entirely.

Also, incidentally speeds up DS_AT() by using the fact that the
stack array is SERIES_FLAG_ALWAYS_DYNAMIC.