Eliminate "Param" datatype trick
With bits in typesets too scarce to spare enough for the parameter
class, the class was snuck into the apparent VAL_TYPE()...leaving the
HEART_BYTE() as a typeset.

Removing the symbol from the typeset opens up enough bits to ditch
that particular hack, making room for more parameter information.

(The ultimate goal is that this be encoded as an ordinary value that
users can reflect and examine, this is only a step toward that.)