Discern NOTE and MARKED flags as persistent or not
The NODE_FLAG_MARKED was being used in some situations where it should
have been a property of the cell, e.g. outlasted overwriting of new
values in that position (e.g. when context variable slots were marked
as being hidden).  But it was also being used in cases where it wasn't
supposed to be sticky, such as marking the output cell of a frame as
stale...which is explicitly intended to be used to detect whether
a new value had been overwritten in that cell or not (without having
to track that flag some other way).

This commit blesses the NODE_FLAG_MARKED bit as the "sticky" one that
is intrinsic to the "format" of the cell itself, while it moves the
non-sticky bits to a CELL_FLAG_NOTE multi-purpose flag.