Fix AUGMENT and make it more efficient
Previously, AUGMENT had to make a new paramlist and a new exemplar
for the augmented function...as well as use a separate dispatcher.

Since paramlists and exemplars are unified, it's now enough to just
make a new paramlist.  And since paramlists are now shared and not
tied into the action's identity, the identity of the action being
augmented can be put into the new paramlist and the same dispatcher
that specialization uses can be used...which doesn't add any new
behavior to a function.