Remove RESKINNED References from Redbol
The removal of RESKINNED means that Redbol can't use it to remove the
constness on the body, getting the Rebol2-style semantics of mutable
function bodies.

But the ability to change parameter conventions is being designed to
come back in a much improved form, under development.  It would be
done through a `f: copy as frame! :func`, making updates to the
parameters in `f`, and then doing `func-nonconst: as action! f`.

In the meantime, Redbol functions won't have the const removal feature.

Also updates some stale usages of LIT.