Get rid of Voidify_Rootparam()
As Make_Action() became the central constructor for actions reused
by all clients, it took on the business of building the archetypal
cell for the action.  However, space needed to be left open for it...
so it was standardized that paramlist-building routines would put
an unreadable void in that slot.  This kept it from being used
prematurely and signaled to the Make_Action() that the caller knew
what they were doing.

Builing paramlists became more complex when they were actually full
contexts, with a series of REBSTR* as the keys that was in parallel
to the parameter information (which also doubled as the exemplar).
Also fewer and fewer routines built their own off-the-cuff paramlists,
choosing instead to define them via natives and reuse them (since
they could now be reused at no cost, since parameters are independent
of the identity of the actions).

In light of these changes, the need for a separate routine to put
the unreadable void doesn't make a lot of sense.  Make_Action() will
complain if the format is wrong, and it's only coming from basically
one origin.

Also cleans out a wasted paramlist made by DOES, which had been
replaced by the reuse of an existing paramlist but just missed and
not thrown out.