Make LINK() and MISC() more ergonomic
The LINK() and MISC() macros originated when REBSER was not a base
class of REBARR or REBSTR/etc.  So they incorporated a cast to
SER() to access the ->link and ->misc fields.

That usefulness vanished with the change to use inheritance to type
check in the C++ build, but to forego checking in the C build (where
REBARR, REBSTR, REBBIN etc. are all just REBSER).

This turns some instances back to referring to ->link and ->misc
directly.  But other instances have something to gain from a macro,
when they are distinct uses that are forced to write to the same
struct field in order to not have problems when read by the GC.
So those cases are rethought in terms of new LINK() and MISC()
macros that coordinate with #defines to give a type-safe LValue and
RValue form for reading and writing that common field...where the
references can be seen uniquely for each purpose.

It also makes REBBMK a subclass of REBSER in the C++ build.