Commit bindings as REBSER, avoid REBNOD aliasing
The concept of the "REBNOD" was to capture a superclass of both
REBSER and REBVAL, which could be pointed to and discerned due to how
they use their header bits.

But due to strict aliasing, updates made through REBVAL->header bits
won't necessarily be seen if you have a REBSER* to the same memory
and try to read the ->header out of it.  Even though the headers are
the same type at the same address, the compiler's logic may think
that the update is not pertinent to refreshing registers if it was
done through different types.

This was a problem, but then there was the REBNOD class, which had
a header that was being used sometimes...even though neither REBVAL
nor REBSER would see those bits.

The issue was understood and hence distinguishing the classes was done
via the first byte in the header--as byte accesses are not subject
to strict aliasing.  The bits used in marking as well as bits to
discern REBSER from REBVAL--or either from UTF-8--all fit in the
first byte.

However, due to having a common name of `header`, mistakes still were
being made; some from early code, and some more recent.

This starts a cleanup effort by renaming the field in the base struct
"leader" instead of header.  It also bites the bullet and commits to
the idea that the binding field does not point to any value cells...
so the binding can be a REBSER, saving time on checking.  Plus a
new BINDING() accessor macro is added for brevity, as `BINDING(x)` is
much shorter than `BINDING(Extra, x).node`.

Since the cast mechanics now catch unnecessary calls to SER() or ARR(),
this improves things across the board, and also allows for the use
of GET_ARRAY_FLAG() in places where it is more convenient--which
improves the readability of the fledgling virtual binding code.