Rethink REBNOD as base class of REBSER and REBVAL
The original concept of a "REBNOD" was to capture pointers that could
be to either REBVAL* or REBSER* variants.  This was based on the
idea that they would share a header format and be superimposable.

As it turned out, strict aliasing makes it unwise to read and write
bits out of pointers to different data types.  You really can't do
it unless they are unions, and REBVAL and REBSER are not.  So the
sharing of header bits became limited to just the contents of the
first byte, when accessed through a byte (which defeats strict
aliasing).

This commit pushes through a more "true" form of commonality for this
REBNOD*, through a base class with no members (in the C++ class) while
the C build of a REBNOD just uses a void*.  Various helpers are
introduced that reduce the need for casting across-the-board.