Name series header "leader" instead of "header"
This continues with the changes oriented around preventing strict
aliasing mistakes, by giving REBVAL* and REBSER* different names for
their header fields.  Even though they are in the same position, they
should not be read and written through distinct pointer types.

In the process of doing the update, sites that would be more clearly
expressed as using the GET_SERIES_FLAG() and GET_ARRAY_FLAG() macros
are changed to use them...as in their modern forms they are efficient
and also help catch bugs due to making sure the INFO vs. FLAG
distinction is honored, since they token past FLAG and INFO on
as part of the macro implementation.