Reclaim SERIES_FLAG_8
The concept behind detecting an END was that it would check the second
byte it received, and see if it was zero.  If it was, then it would
assume the pointer was to an END.

For cells, this fit nicely with the idea of the type slot being 0.

For the series "info" header, it fit with arrays saying their WIDE
byte was 0 (since width access was hardcoded at the REBVAL size at
all the reference sites, that used ARR_AT())

However, this needed some means of distinguishing a cell "leader".
It was possible for the second byte of such series to be 0.  As an
efficiency trick, a flag was reserved in the second byte to always
be 1, to distinguish the situation.

But SERIES_FLAG_XXX are scarce, and giving up this bit isn't ideal.
Nor is sacrificing a bit in the extremely scarce first byte.

This switches it so that the series info bits have the bit set claiming
that they represent a cell, so that the END illusion is done with
that flag.  This has a downside because it gives the impression that
there's a full size cell slot behind that END when there is not.  But
termination will not ultimately be done with this mechanism, and the
better plan will be to remove the NODE_FLAG_NODE and trust that out
of bounds access past the tail will crash on that, when it's no
longer necessary to represent ENDs in the info spot.