Automatic array termination, binary may not terminate
This is a starting step toward having arrays not be terminated with
END (except in debug builds, in order to catch overruns more easily).

Hence termination is removed as a concern from most high-level
callsites, instead moved to SET_SERIES_USED().

Previously that low-level operation would set the length of strings
(different from bytes used) to trash in debug builds, to make sure
higher level operations would update it.  This takes that concept
a bit further, by having debug builds poison the tail byte of series
which are byte sized.  This forces clients who care (e.g. strings)
to do an update of it, while clients who don't care (binaries never
aliased as strings) wouldn't necessarily have to terminate.

This essentially removes the concept of generalized termination, e.g.
memset() of an arbitrary-sized element at the tail to zero bytes...
which was not used by keylists, and is soon to not be used by arrays.