Use data stack instead of BUF_COLLECT
Collecting words was a shared piece of functionality used by both
COLLECT-WORDS the native, and Make_Object() in order to blend the
noticing of duplicate keys with having a binding table for new keys
being added.

The collection process needs to remember the words to remove them
from the binder when done.  This was previously done in a BUF_COLLECT
which was a kind of slow and redundant version of the data stack.  It
was kept because there was some feeling that it would need to be
consulted in case of a failure during a collection--and the data
stack would not be persistent enough for this purpose.

But there was no failure tolerance and the code had to be engineered to
not fail during a collect (as with anything else that has a binder in
effect), so removing the collect buffer and using the data stack is an
improvement.  It means that there isn't redundant code for managing
expansion and delayed array termination--the data stack does this
already.